contentbotapi/requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
pydantic
